@@ -1,72 +1,69 @@
 {
-  "name": "laravel/laravel",
-  "description": "The Laravel Framework.",
-  "keywords": [
-    "framework",
-    "laravel"
-  ],
-  "license": "MIT",
-  "type": "project",
-  "require": {
-    "php": ">=5.5.9",
-    "laravel/framework": "5.1.*",
-    "rcrowe/twigbridge": "0.8.x",
-    "stolz/assets": "^0.1.4",
-    "laravelcollective/html": "5.1.*",
-    "components/jquery": "^1.11.2",
-    "twbs/bootstrap": "^3.3.5",
-    "eonasdan/bootstrap-datetimepicker": "^4.17",
-    "fortawesome/font-awesome": "^4.4",
-    "afarkas/html5shiv": "^3.7.3",
-    "rogeriopradoj/respond": "^1.4.2",
-    "datatables/datatables": "^1.10",
-    "spatie/laravel-tail": "^1.1.1",
-    "roave/security-advisories": "dev-master"
-  },
-  "require-dev": {
-    "fzaninotto/faker": "~1.4",
-    "mockery/mockery": "0.9.*",
-    "phpunit/phpunit": "~4.0",
-    "phpspec/phpspec": "~2.1",
-    "barryvdh/laravel-debugbar": "~2.0"
-  },
-  "autoload": {
-    "classmap": [
-      "database"
-    ],
-    "psr-4": {
-      "SQLgreyGUI\\": "app/"
+    "name": "laravel/laravel",
+    "description": "The Laravel Framework.",
+    "keywords": ["framework", "laravel"],
+    "license": "MIT",
+    "type": "project",
+    "require": {
+        "php": ">=5.5.9",
+        "laravel/framework": "5.2.*",
+        "roave/security-advisories": "dev-master",
+        "doctrine/dbal": "^2.5",
+        "rcrowe/twigbridge": "^0.9.2",
+        "lbausch/laravel-cornerstone": "^3.0",
+        "stolz/assets": "^0.1.4",
+        "laravelcollective/html": "5.2.*",
+        "spatie/laravel-tail": "^1.1.1",
+        "components/jquery": "^2.2.0",
+        "twbs/bootstrap": "^3.3.5",
+        "eonasdan/bootstrap-datetimepicker": "^4.17",
+        "fortawesome/font-awesome": "^4.4",
+        "afarkas/html5shiv": "^3.7.3",
+        "rogeriopradoj/respond": "^1.4.2",
+        "datatables/datatables": "^1.10",
+        "onokumus/metismenu": "^2.4"
+    },
+    "require-dev": {
+        "fzaninotto/faker": "~1.4",
+        "mockery/mockery": "0.9.*",
+        "phpunit/phpunit": "~4.0",
+        "symfony/css-selector": "2.8.*|3.0.*",
+        "symfony/dom-crawler": "2.8.*|3.0.*",
+        "laravel/homestead": "^3.0",
+        "barryvdh/laravel-debugbar": "~2.0",
+        "barryvdh/laravel-ide-helper": "^2.1"
+    },
+    "autoload": {
+        "classmap": [
+            "database"
+        ],
+        "psr-4": {
+            "SQLgreyGUI\\": "app/"
+        }
+    },
+    "autoload-dev": {
+        "classmap": [
+            "tests/TestCase.php"
+        ]
+    },
+    "scripts": {
+        "post-root-package-install": [
+            "php -r \"copy('.env.example', '.env');\""
+        ],
+        "post-create-project-cmd": [
+            "php artisan key:generate"
+        ],
+        "post-install-cmd": [
+            "php artisan clear-compiled",
+            "php artisan optimize"
+        ],
+        "post-update-cmd": [
+            "php artisan clear-compiled",
+            "php artisan optimize"
+        ]
+    },
+    "config": {
+        "preferred-install": "dist",
+        "component-dir": "vendor"
     }
-  },
-  "autoload-dev": {
-    "classmap": [
-      "tests/TestCase.php"
-    ]
-  },
-  "scripts": {
-    "post-install-cmd": [
-      "php artisan clear-compiled",
-      "php artisan optimize"
-    ],
-    "pre-update-cmd": [
-      "php artisan clear-compiled"
-    ],
-    "post-update-cmd": [
-      "php artisan optimize",
-      "php artisan route:clear",
-      "php artisan config:cache",
-      "php artisan asset:flush --force",
-      "php artisan twig:clean"
-    ],
-    "post-root-package-install": [
-      "php -r \"copy('.env.example', '.env');\""
-    ],
-    "post-create-project-cmd": [
-      "php artisan key:generate"
-    ]
-  },
-  "config": {
-    "preferred-install": "dist",
-    "component-dir": "vendor"
-  }
 }
