@@ -3,44 +3,12 @@
 namespace SQLgreyGUI\Http\Controllers;
 
 use Illuminate\Foundation\Bus\DispatchesJobs;
-use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use Auth;
-use Request;
 
-abstract class Controller extends BaseController
+abstract class Controller extends \Bausch\LaravelCornerstone\Http\Controllers\CornerstoneController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
-
-    /**
-     * The user id.
-     *
-     * @var int
-     */
-    protected $userid;
-
-    /**
-     * User.
-     *
-     * @var \SQLgreyGUI\Models\User
-     */
-    protected $user;
-
-    /**
-     * Constructor.
-     */
-    public function __construct()
-    {
-        if (Auth::check()) {
-            $this->user = Auth::User();
-
-            $this->userid = $this->user->getId();
-
-            // pass the user object to all views
-            view()->share('user', $this->user);
-        }
-    }
 
     /**
      * Parse provided entries.
@@ -49,9 +17,9 @@ abstract class Controller extends BaseController
      */
     protected function parseEntries($input_identifier, $repository)
     {
-        $items_tmp = Request::input($input_identifier, array());
+        $items_tmp = request()->input($input_identifier, []);
 
-        $items = array();
+        $items = [];
 
         $repository = app($repository);
 
@@ -65,36 +33,12 @@ abstract class Controller extends BaseController
     }
 
     /**
-     * Get name of the current controller.
-     *
-     * @return string
-     */
-    protected function getController()
-    {
-        $controller = explode('\\', strtok(\Route::currentRouteAction(), '@'));
-
-        return end($controller);
-    }
-
-    /**
-     * Get action.
-     *
-     * @param string $action
-     *
-     * @return string
-     */
-    public function getAction($action)
-    {
-        return $this->getController().'@'.$action;
-    }
-
-    /**
      * Is ajax request?
      *
      * @return bool
      */
     protected function isAjax()
     {
-        return Request::ajax();
+        return request()->ajax();
     }
 }
