@@ -13,9 +13,9 @@ class EventServiceProvider extends ServiceProvider
      * @var array
      */
     protected $listen = [
-        'App\Events\SomeEvent' => [
-            'App\Listeners\EventListener',
-        ],
+        //'App\Events\SomeEvent' => [
+        //    'App\Listeners\EventListener',
+        //],
     ];
 
     /**
