@@ -4,38 +4,36 @@ namespace SQLgreyGUI\Repositories;
 
 use SQLgreyGUI\Models\AwlEmail as Email;
 
-class AwlEmailRepositoryEloquent implements AwlEmailRepositoryInterface
+class AwlEmailRepositoryEloquent extends BaseRepositoryEloquent implements AwlEmailRepositoryInterface
 {
+    public function __construct(Email $email)
+    {
+        $this->model = $email;
+    }
 
     public function findAll()
     {
-        $data = Email::orderBy('sender_name', 'asc')->get();
+        $data = $this->model->orderBy('sender_name', 'asc')->get();
 
         return $data;
     }
 
-    public function instance($data = array())
+    public function store($email)
     {
-        return new Email($data);
-    }
-
-    public function store(Email $email)
-    {
-        return Email::insert(array(
-                    'sender_name' => $email->getSenderName(),
-                    'sender_domain' => $email->getSenderDomain(),
-                    'src' => $email->getSource(),
-                    'first_seen' => $email->getFirstSeen(),
-                    'last_seen' => $email->getLastSeen(),
+        return $this->model->insert(array(
+            'sender_name' => $email->getSenderName(),
+            'sender_domain' => $email->getSenderDomain(),
+            'src' => $email->getSource(),
+            'first_seen' => $email->getFirstSeen(),
+            'last_seen' => $email->getLastSeen(),
         ));
     }
 
-    public function destroy(Email $email)
+    public function destroy($email)
     {
-        return Email::where('sender_name', $email->getSenderName())
-                        ->where('sender_domain', $email->getSenderDomain())
-                        ->where('src', $email->getSource())
-                        ->delete();
+        return $this->model->where('sender_name', $email->getSenderName())
+            ->where('sender_domain', $email->getSenderDomain())
+            ->where('src', $email->getSource())
+            ->delete();
     }
-
 }
