@@ -3,40 +3,25 @@
 namespace SQLgreyGUI\Repositories;
 
 use Carbon\Carbon;
-use SQLgreyGUI\Models\Greylist;
 
 interface GreylistRepositoryInterface extends BaseRepositoryInterface
 {
-
     /**
-     * find entries for given period of time
-     * 
-     * @param string $period
+     * If $end is omitted, the current timestamp is used.
+     *
+     * @param Carbon      $start
+     * @param Carbon|null $end
+     *
+     * @return \Illuminate\Database\Eloquent\Collection
      */
-    public function findByPeriod($period = null);
+    public function findBetween(Carbon $start, Carbon $end = null);
 
     /**
-     * if $end is omitted, the current timestamp is used
-     * 
-     * @param int $start
-     * @param int $end
-     */
-    public function findBetween($start, $end = null);
-
-    /**
-     * delete
-     * 
-     * @param Greylist $greylist
-     * 
-     * @return bool
-     */
-    public function destroy(Greylist $greylist);
-
-    /**
-     * destroy entries older than given date
-     * 
+     * Destroy entries older than given date.
+     *
      * @param Carbon $date
-     * @return int
+     *
+     * @return bool
      */
     public function destroyOlderThan(Carbon $date);
 }
