@@ -4,47 +4,27 @@ namespace SQLgreyGUI\Repositories;
 
 trait OptDomainTrait
 {
-
-    private $model_class;
-
-    /**
-     * Constructor
-     */
     public function __construct()
     {
-        $this->model_class = 'SQLgreyGUI\Models\\' . $this->model;
+        $this->model = app('SQLgreyGUI\Models\\'.$this->model_class);
     }
 
     public function findAll()
     {
-        $data = new $this->model_class;
-
-        $data = $data::orderBy('domain', 'asc')->get();
+        $data = $this->model->orderBy('domain', 'asc')->get();
 
         return $data;
     }
 
-    public function instance($data = array())
-    {
-        $model = $this->model_class;
-
-        return new $model($data);
-    }
-
     public function destroy($domain)
     {
-        $model = $this->model_class;
-
-        return $model::where('domain', $domain->getDomain())->delete();
+        return $this->model->where('domain', $domain->getDomain())->delete();
     }
 
     public function store($domain)
     {
-        $model = $this->model_class;
-
-        return $model::insert(array(
-                    'domain' => $domain->getDomain(),
-        ));
+        return $this->model->insert([
+            'domain' => $domain->getDomain(),
+        ]);
     }
-
 }
