@@ -2,20 +2,7 @@
 
 namespace SQLgreyGUI\Repositories;
 
-
-interface OptEmailRepositoryInterface extends BaseRepositoryInterface {
-
-    /**
-     * destroy email
-     * 
-     * @param Email $email
-     */
-    public function destroy($email);
-
-    /**
-     * store email
-     * 
-     * @param Email $email
-     */
-    public function store($email);
+interface OptEmailRepositoryInterface extends BaseRepositoryInterface
+{
+    //
 }
