@@ -4,47 +4,27 @@ namespace SQLgreyGUI\Repositories;
 
 trait OptEmailTrait
 {
-
-    private $model_class;
-
-    /**
-     * Constructor
-     */
     public function __construct()
     {
-        $this->model_class = 'SQLgreyGUI\Models\\' . $this->model;
+        $this->model = app('SQLgreyGUI\Models\\'.$this->model_class);
     }
 
     public function findAll()
     {
-        $model = new $this->model_class;
-
-        $data = $model::orderBy('email', 'asc')->get();
+        $data = $this->model->orderBy('email', 'asc')->get();
 
         return $data;
     }
 
-    public function instance($data = array())
-    {
-        $model = $this->model_class;
-
-        return new $model($data);
-    }
-
     public function destroy($email)
     {
-        $model = $this->model_class;
-
-        return $model::where('email', $email->getEmail())->delete();
+        return $this->model->where('email', $email->getEmail())->delete();
     }
 
     public function store($email)
     {
-        $model = $this->model_class;
-
-        return $model::insert(array(
-                    'email' => $email->getEmail(),
-        ));
+        return $this->model->insert([
+            'email' => $email->getEmail(),
+        ]);
     }
-
 }
