@@ -4,43 +4,10 @@ namespace SQLgreyGUI\Repositories;
 
 use SQLgreyGUI\Models\User as User;
 
-class UserRepositoryEloquent implements UserRepositoryInterface
+class UserRepositoryEloquent extends BaseRepositoryEloquent implements UserRepositoryInterface
 {
-
-    public function findById($id)
+    public function __construct(User $user)
     {
-        $data = User::where('id', $id);
-
-        if ($data->count() > 0) {
-            return $data->first();
-        }
+        $this->model = $user;
     }
-
-    public function findAll()
-    {
-        $data = User::all();
-
-        return $data;
-    }
-
-    public function instance($data = array())
-    {
-        return new User($data);
-    }
-
-    public function store(User $user)
-    {
-        return $user->save();
-    }
-
-    public function update(User $user)
-    {
-        return $user->save();
-    }
-
-    public function destroy(User $user)
-    {
-        return $user->delete();
-    }
-
 }
