@@ -49,12 +49,19 @@ return [
         'sqlite' => [
             'driver' => 'sqlite',
             'database' => database_path('database.sqlite'),
+            'prefix' => env('DB_PREFIX', 'sg_'),
+        ],
+
+        'sqlite_sqlgrey' => [
+            'driver' => 'sqlite',
+            'database' => database_path('database_sqlgrey.sqlite'),
             'prefix' => '',
         ],
 
         'mysql' => [
             'driver' => 'mysql',
             'host' => env('DB_HOST', 'localhost'),
+            'port' => env('DB_PORT', '3306'),
             'database' => env('DB_DATABASE'),
             'username' => env('DB_USERNAME'),
             'password' => env('DB_PASSWORD'),
@@ -62,11 +69,13 @@ return [
             'collation' => 'utf8_unicode_ci',
             'prefix' => env('DB_PREFIX', 'sg_'),
             'strict' => false,
+            'engine' => null,
         ],
 
-        'sqlgrey' => [
+        'mysql_sqlgrey' => [
             'driver' => 'mysql',
             'host' => env('SQLGREY_DB_HOST', env('DB_HOST')),
+            'port' => env('SQLGREY_DB_PORT', env('DB_PORT')),
             'database' => env('SQLGREY_DB_DATABASE', env('DB_DATABASE')),
             'username' => env('SQLGREY_DB_USERNAME', env('DB_USERNAME')),
             'password' => env('SQLGREY_DB_PASSWORD', env('DB_PASSWORD')),
@@ -74,27 +83,19 @@ return [
             'collation' => 'utf8_unicode_ci',
             'prefix' => '',
             'strict' => false,
+            'engine' => null,
         ],
 
         'pgsql' => [
             'driver' => 'pgsql',
             'host' => env('DB_HOST', 'localhost'),
+            'port' => env('DB_PORT', '5432'),
             'database' => env('DB_DATABASE', 'forge'),
             'username' => env('DB_USERNAME', 'forge'),
             'password' => env('DB_PASSWORD', ''),
             'charset' => 'utf8',
             'prefix' => '',
             'schema' => 'public',
-        ],
-
-        'sqlsrv' => [
-            'driver' => 'sqlsrv',
-            'host' => env('DB_HOST', 'localhost'),
-            'database' => env('DB_DATABASE', 'forge'),
-            'username' => env('DB_USERNAME', 'forge'),
-            'password' => env('DB_PASSWORD', ''),
-            'charset' => 'utf8',
-            'prefix' => '',
         ],
 
     ],
@@ -128,8 +129,9 @@ return [
         'cluster' => false,
 
         'default' => [
-            'host' => '127.0.0.1',
-            'port' => 6379,
+            'host' => env('REDIS_HOST', 'localhost'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', 6379),
             'database' => 0,
         ],
 
